+++ Expanded Cheat Menu/Code/Menus/CheatsMenu.lua
@@ -39,7 +39,7 @@ Actions[c] = {ActionName = TranslationTable[11412--[[Trigger fireworks]]],
 	ActionId = ".Trigger fireworks",
 	ActionIcon = "CommonAssets/UI/Menu/DisableRMMaps.tga",
 	RolloverText = TranslationTable[302535920001402--[[Add some party to your domes for 3 hours (10 domes max).]]],
-	OnAction = TriggerFireworks,
+	OnAction = ChoGGi.MenuFuncs.TriggerFireworks,
 }
 
 c = c + 1