+++ Mods ChoGGi/Allow BB Buildings/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2595725610",
 	"pops_any_uuid", "72724cbe-731f-4edd-b7bb-fcabaa1b10ec",
 	"lua_revision", 1007000, -- Picard
-	"version", 6,
+	"version", 8,
 	"version_major", 0,
-	"version_minor", 6,
+	"version_minor", 8,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {