+++ Mods ChoGGi/RC Garage/metadata.lua
@@ -11,9 +11,9 @@ return PlaceObj("ModDef", {
 	"id", "ChoGGi_RCGarage",
 	"steam_id", "1557866331",
 	"pops_any_uuid", "fd81692a-a91f-4f22-94f9-ad24a9442e01",
-	"version", 11,
+	"version", 12,
 	"version_major", 1,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"lua_revision", 1007000, -- Picard